+++ lib/rspec/mocks/any_instance/message.rb
@@ -1,0 +1,13 @@
+module RSpec
+  module Mocks
+    module AnyInstance
+
+      class Message < Struct.new(:name, :args, :block)
+        def unplayable?(arguments)
+          name == :with && args != arguments
+        end
+      end
+    
+    end
+  end
+end